+++ docs_manual/conf.py
@@ -15,6 +15,8 @@
 import os
 import sys
 
+import digestiflow
+
 sys.path.insert(0, os.path.abspath("."))
 
 # Get the project root dir, which is the parent dir of this
@@ -25,8 +27,6 @@ project_root = os.path.dirname(cwd)
 # This lets us ensure that the source package is imported, and that its
 # version is used.
 sys.path.insert(0, project_root)
-
-import digestiflow
 
 
 # -- Project information -----------------------------------------------------