+++ HISTORY.rst
@@ -11,6 +11,7 @@ HEAD (unreleased)
 - Paginating large lists.
 - More prefetching for large list for speedups.
 - Fixing organism/reference editing.
+- More refined checking of barcodes and demultiplexing cycles.
 
 ------
 v1.0.0